requirements.txt
Flask
Flask-Cors
Flask-Login
Flask-Bcrypt
Flask-Caching
Flask-SQLAlchemy
Flask-WTF
redis
black
pydantic
gunicorn
python-dotenv
email-validator
Werkzeug
rpi_ws281x
Adafruit-Blinka==8.39.1
adafruit-circuitpython-neopixel==6.3.11
Flask==3.0.3
pillow
